pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "ddos_deny_list"
version = "1.0.0"
description = "SilverLine Deny-list project"
authors = ["Network to Code <info@networktocode.com>"]
license = "PROPRIETARY"

[tool.poetry.dependencies]
python = "^3.6"
requests = "^2.25.1"
ipaddress = "^1.0.23"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
pytest = "^5.4.1"
pytest-mock = "^3.3.1"
pytest-cov = "^2.10.1"
requests_mock = "^1.7.0"
pyyaml = "^5.3"
black = "^19.10b0"
pylint = "^2.4.4"
pydocstyle = "^5.0.2"
yamllint = "^1.20.0"
bandit = "^1.6.2"
invoke = "^1.4.1"
flake8 = "^3.8.3"

[tool.pytest.ini_options]
testpaths = [
    "tests"
]

[tool.pylint.messages_control]
# Line length is enforced by Black, so pylint doesn't need to check it.
# Pylint and Black disagree about how to format multi-line arrays; Black wins.
disable = """,
    line-too-long,
    bad-continuation,
    """

[tool.pylint.miscellaneous]
# Don't flag TODO as a failure, let us commit with things that still need to be done in the code
notes = """,
    FIXME,
    XXX,
    """

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
        /(
          \.git
        | \.tox
        | \.venv
        | env/
        | _build
        | build
        | dist
        )/
    '''

